add Jenkinsfile

+++ Jenkinsfile
@@ -35,25 +35,17 @@ pipeline {
 
                     echo "⏹️ Eski jarni backup va to‘xtatish..."
                     sh """
-                    ssh -o StrictHostKeyChecking=no ${DEPLOY_SERVER} '
-                        mkdir -p ${BACKUP_PATH}
-                        if [ -f ${DEPLOY_PATH}${JAR_NAME} ]; then
-                            mv ${DEPLOY_PATH}${JAR_NAME} ${BACKUP_PATH}${JAR_NAME}_\$(date +%Y%m%d%H%M%S)
-                        fi
-                        pkill -f "${JAR_NAME}" || true
-                    '
+                    ssh -o StrictHostKeyChecking=no ${DEPLOY_SERVER} 'mkdir -p ${BACKUP_PATH} && \
+                    if [ -f ${DEPLOY_PATH}${JAR_NAME} ]; then mv ${DEPLOY_PATH}${JAR_NAME} ${BACKUP_PATH}${JAR_NAME}_\$(date +%Y%m%d%H%M%S); fi && \
+                    pkill -f "${JAR_NAME}" || true'
                     """
 
                     echo "▶️ Yangi jarni ishga tushirish..."
-                    sh """
-                    ssh -o StrictHostKeyChecking=no ${DEPLOY_SERVER} '
-                        nohup java -jar ${DEPLOY_PATH}${JAR_NAME} > ${LOG_FILE} 2>&1 &
-                        disown
-                    '
-                    """
+                    sh "ssh -o StrictHostKeyChecking=no ${DEPLOY_SERVER} 'nohup java -jar ${DEPLOY_PATH}${JAR_NAME} > ${LOG_FILE} 2>&1 &'"
                 }
             }
         }
+
 
 
     }
